refactor : 코치 이름에 대한 검증을 Coache 도메인 객체에서 수행하도록 변경
