+++ src/main/java/menu/view/InputView.java
@@ -2,14 +2,14 @@ package menu.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import menu.util.StringConvertor;
-import menu.view.validation.CoacheInputValidator;
+import menu.view.validation.InputValidator;
 
 public class InputView {
 
     public String inputCoaches() {
         String input = Console.readLine();
         String trimedInput = StringConvertor.convertTrimedString(input);
-        CoacheInputValidator.validate(trimedInput);
+        InputValidator.validateSeparator(trimedInput);
         return trimedInput;
     }
 }
